🔨 Refactor: rename dir, refactory slack class
