+++ bolt_python/actions.py
@@ -4,9 +4,9 @@ from logging import Logger
 from slack_bolt import BoltContext
 from slack_sdk import WebClient
 
-from bolt.app import bolt_app
-from bolt.forms import modal_form
-from bolt.utils import validate_reservation, get_random, datetime_to_timestamp
+from bolt_python.app import bolt_app
+from bolt_python.forms import modal_form
+from bolt_python.utils import validate_reservation, get_random, datetime_to_timestamp
 from notion.actions import create_reservation, get_lunch
 
 
@@ -273,4 +273,4 @@ def send_dm(ack, body, client, view, logger):
         logger.exception(f"발송실패 {e}")
 
 
-actions_check = "bolt action is called!"  # instead handler.start() | call check
+actions_check = "bolt_python action is called!"  # instead handler.start() | call check
